changes to improve the demo

+++ src/main/java/org/superbiz/moviefun/MoviesBean.java
@@ -51,7 +51,30 @@ public class MoviesBean {
 
     public void deleteMovie(long id) {
         Movie movie = entityManager.find(Movie.class, id);
+        if (movie == null) {
+            throw new IllegalArgumentException("Movie " + id + " not found");
+        }
         entityManager.remove(movie);
+    }
+
+    public Movie addCommentToMovie(final Long id, final Comment comment) {
+        final Movie movie = entityManager.find(Movie.class, id);
+        if (movie == null) {
+            throw new IllegalArgumentException("Movie " + id + " not found");
+        }
+        entityManager.persist(comment);
+        movie.getComments().add(comment);
+        return movie;
+    }
+
+    public Movie removeCommentToMovie(final Long id, final Comment comment) {
+        final Movie movie = entityManager.find(Movie.class, id);
+        if (movie == null) {
+            throw new IllegalArgumentException("Movie " + id + " not found");
+        }
+        movie.getComments().remove(comment);
+        entityManager.remove(comment);
+        return movie;
     }
 
     public List<Movie> getMovies(Integer firstResult, Integer maxResults, String field, String searchTerm) {
